StyleCop: SA1402
